pyproject.toml
[tool.poetry]
name = "autoapply"
version = "0.1.0"
description = "Automatically apply to jobs"
authors = ["Amir Braham <amirbrahamm@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
selenium = "^4.27.1"
webdriver-manager = "^4.0.2"
langchain = "^0.3.13"
langchain-mistralai = "^0.2.4"
mistralai = "^1.2.5"
playwright = "^1.49.1"
langgraph = "^0.2.60"
langchain-community = "^0.3.13"
pillow = "^11.1.0"
nest-asyncio = "^1.6.0"
beautifulsoup4 = "^4.12.3"
lxml = "^5.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
